requirements.txt
selenium==4.33.0
beautifulsoup4
python-dotenv
requests